Fix a couple bugs in checking FreeMaps of defthms.
One bug went only undetected because convert.js creates FreeMaps as sparse
arrays; but if you converted them to JSON and back, verification would fail. We
now do this for every theorem and validate before and after. This was a
false-negative bug for any defthm whose binding arg was not in the first
position.

The other bug changed the way that we validate the FreeMap of the new term. This
could potentially have caused a false-positive validation for some sufficiently
bizarre facts. (I currently have only a limited amount of confidence in this
code; it could use some more tests.)